Move turn factory to its own file
